requirements.txt
amqp==5.0.2
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
celery-progress==0.0.14
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
dj-stripe==2.4.1
Django==3.1.5
django-allauth==0.44.0
djangorestframework==3.12.2
idna==2.10
iexfinance==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
kombu==5.0.2
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
pandas==1.2.0
plotly==4.14.1
prompt-toolkit==3.0.8
pycparser==2.20
PyJWT==2.0.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
requests-cache==0.5.2
requests-oauthlib==1.3.0
retrying==1.3.3
sentry-sdk==0.19.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
stripe==2.55.1
tweepy==3.10.0
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
